Change how we configure MKL single-dynamic library
This is mostly beneficial for building the Conda package.

Previously, a header was generated at compile-time with the values of
interface and threading layer appropriate for the given compiler.
This change should un-tie us from one specific BLAS/LAPACK
implementation at compile-time: it should be now possible to build a
"generic" conda package and then switch
Netlib/OpenBLAS/Accelerate/BLIS/MKL when installing.