Fix number of threads used in GPU kernel launch code
The code in FockDriverGPU.hip declared OpenMP regions for the maximum
number of threads available, but then only ever used one thread per
available device to allocate the data and launch the GPU kernel.

This has been changed now to only ever use one thread per device, which
simplifies the code and makes it easier to rationalize what is happening
on the host.

Prepares for using global allocations of static data instead of
re-allocating everything for each function call.